Target_state refactored to support modules
The class Target_state does not directly hold the stored data of a checkpointer
anymore. Instead every module provides a container class, which stores the
module-pending data. Such container class need to be derived from the class
Module_state.

When a module is checkpointed, its customized Module_state is stored in the
Target_state.

When a module should be restored, the Target_state returns its customized
Module_state based on the name of the module. Finally the Module_state is handed
over by the `restore(Module_state *state)` method of the module.

This concept requires two major changes:
1. Every module holds its session state in the module class
2. Every module creates its own Module_state object.

In order to simplify the usage of `Genode::Allocator`s, then Module_state object
is created with the same allocator as every other heap object.