implement safe list element handling
The current approach suffers by two problems. One list element can not be added to
two lists, as the element itself stores the linking information. Therefore a
fifo is used to manage destroyed elements. This has the advantage, that each
element also provide the method enqueued() which helps to conclude if the
element is already destroyed.

The second problem is about optimizing the ck_* and _new* lists. Instead of transfering
elements from _new* to ck_* during checkpoint, all elements are stored in one
list named _*. The ck_* just points to an element in the list, which was the
head during checkpoint.