replace Service::Factory by Root_components again
A Service factory can not be used to announce a service. Therefore the
previously used Root_components are still necessary. These are implemented again
and refactored as template class Rtcr::Root_component<SESSION> in
rtcr/root_component.h. The additionally Service::Factory is not necessary, as a
service can be created from a root component.

Due to the fact, that the PD and CPU session are initialized based on the
Child_policy::resolve_session_request(...), the child does not depend on the
RTCR implementation anymore. The exchange of the intercepting services between
RTCR module and child is managed by the Genode::Registry<Genode::Service> in
future.