Replace custom child implementation by custom init process
The Rtcr::Child implementation is not fully ported to Genode 18.02. Due an
unknown error during child creation, the child component seems to freeze. As the
Rtcr::Child is just a simple version of the Init Process, the Init Process is
forked and used as replacement.

The Init Process (`init`) is named `rtcr_init` and comes with an intergrated
Rtcr. A typical routing of a Rtcr child looks like that:

```
<route>
    <service name="Timer"> <rtcr label="sheep_counter"/> </service>
    <service name="ROM">  <parent/> </service>
    <any-service> <rtcr/> </any-service>
</route>
```
1. As the Timer has a bug and choose a wrong label, the label need to be fixed
by setting it to the child application label.

2. A child has multiple ROM session, which lead to problems to the current
Rtcr. Therefore all ROM sessions are routed directly to the parent.

3. All other sessions are routed to the RTCR component, by using the `<rtcr/>`
node.

The `rtcr_init` also announces a `Rtcr::Session` service, which provides an
interface to checkpoint all rtcr children and transfer a shared memory dataspace
to the caller. The dataspace contains the compressed checkpoint of all children.