integrate Genode Report Session API for publishing internal Rtcr timestamps
This changes allows to forward the timestamps of each checkpointable's
checkpoint to a Report Rom. Other applications are able to read this values.

This requires several adjustments in a run script:

1. Add `server/report_rom` which holds the values in a dataspaces.
2. Enable the report function in `rtcr_init`. The node `<reporter/>` enables the
report functions of the default init component. The rtcr_state can be enabled by
enabling the attribut `rtcr: `<reporter rtcr="true"/>`.
3. Each component which reads from/writes to the report rom requires the routing
items `Report` and `ROM`.
4. The current values of the report rom are shown, if `verbose="yes` is set in
the `<config>` node of `<start name="report_rom"/>.